Include the filtering php file from build/ folder instead of src/
- Removed `render.php` and replaced it with `filters.php` to handle mobile section headers.
- Updated webpack configuration to copy `filters.php` to the build folder.
- Added functionality to dynamically create mobile headers for group blocks with sticky menu enabled.
- Updated changelog to reflect changes in sticky menu block functionality.
- Adjusted CSS class audit documentation to reference new filters file.
