+++ package.json
@@ -42,6 +42,7 @@
     "@wordpress/prettier-config": "4.37.0",
     "@wordpress/scripts": "^31.2.0",
     "@wordpress/stylelint-config": "^23.29.0",
+    "copy-webpack-plugin": "^10.2.4",
     "eslint": "9.39.2",
     "jest": "30.2.0",
     "npm-package-json-lint": "9.1.0",
